Simplified return value of EthernetFrame.getLengthInBytes(), now without static cast

+++ src/kernel/network/ethernet/EthernetFrame.h
@@ -29,7 +29,7 @@ public:
 
     uint8_t copyTo(Kernel::NetworkByteBlock *output);
 
-    uint16_t getLengthInBytes();
+    size_t getLengthInBytes();
 
     void setSourceAddress(EthernetAddress *source);
 };
